move example function to its own file
